Add some simple index to enhance completion speed.
Completion requests had conditions using 'deleted' field on each
table and not using any indexes because of use of LIKE or ILIKE. We
don't build indexes compatible with LIKE/ILIKE but we add an index on
'deleted' field. Most results will be trimmed down and the completion
request can be completed in far less time than without index.